requirements.txt
Django==2.2.3
Jinja2==2.10.1
MarkupSafe==1.1.1
pytz==2019.1
PyYAML==5.1.1
sqlparse==0.3.0
